Fix for code highlights
The code is now rendered as a block and then set mutliple times, once for every line. The code is then moved and clipped to show only the necessary line in one table cell.
THis approach works for now (though moving the code requires some measuring to get the displacement right), but might break for some settings.

Drawbacks are, that linebreaks in long lines are not rendered correctly for now and code selection in PDF-Editors does not work, since they select the (hidden) code multiple times.